Rename packet -> ingest
