+++ zqd/handlers.go
@@ -13,7 +13,7 @@ import (
 	"github.com/brimsec/zq/pcap/pcapio"
 	"github.com/brimsec/zq/pkg/nano"
 	"github.com/brimsec/zq/zqd/api"
-	"github.com/brimsec/zq/zqd/packet"
+	"github.com/brimsec/zq/zqd/ingest"
 	"github.com/brimsec/zq/zqd/search"
 	"github.com/brimsec/zq/zqd/space"
 	"github.com/gorilla/mux"
@@ -65,11 +65,11 @@ func handlePacketSearch(c *Core, w http.ResponseWriter, r *http.Request) {
 	if err := req.FromQuery(r.URL.Query()); err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
 	}
-	if s.PacketPath() == "" || !s.HasFile(packet.IndexFile) {
+	if s.PacketPath() == "" || !s.HasFile(ingest.IndexFile) {
 		http.Error(w, "space has no pcaps", http.StatusNotFound)
 		return
 	}
-	index, err := pcap.LoadIndex(s.DataPath(packet.IndexFile))
+	index, err := pcap.LoadIndex(s.DataPath(ingest.IndexFile))
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -148,7 +148,7 @@ func handleSpaceGet(c *Core, w http.ResponseWriter, r *http.Request) {
 	}
 	info := &api.SpaceInfo{
 		Name:          s.Name(),
-		PacketSupport: s.HasFile(packet.IndexFile),
+		PacketSupport: s.HasFile(ingest.IndexFile),
 		PacketPath:    s.PacketPath(),
 	}
 	if s.HasFile("all.bzng") {
