+++ zqd/handlers_zeek_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/brimsec/zq/pkg/test"
 	"github.com/brimsec/zq/zqd"
 	"github.com/brimsec/zq/zqd/api"
-	"github.com/brimsec/zq/zqd/packet"
+	"github.com/brimsec/zq/zqd/ingest"
 	"github.com/brimsec/zq/zqd/zeek"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -67,7 +67,7 @@ func TestPacketPostSuccess(t *testing.T) {
 		assert.Equal(t, p.pcapfile, info.PacketPath)
 	})
 	t.Run("PacketIndexExists", func(t *testing.T) {
-		require.FileExists(t, filepath.Join(p.core.Root, p.space, packet.IndexFile))
+		require.FileExists(t, filepath.Join(p.core.Root, p.space, ingest.IndexFile))
 	})
 	t.Run("TaskStartMessage", func(t *testing.T) {
 		status := p.payloads[0].(*api.TaskStart)
